PoC of storage optimization for substakes: use uint96 to store NU amounts
Optimization works up to uint208; bigger than that, the substake struct is back at 2 slots again

+++ nucypher/blockchain/eth/sol/source/contracts/StakingEscrow.sol
@@ -64,7 +64,7 @@ contract StakingEscrow is Issuer {
         uint16 firstPeriod;
         uint16 lastPeriod;
         uint16 periods;
-        uint256 lockedValue;
+        uint96 lockedValue;
     }
 
     struct Downtime {
@@ -584,7 +584,7 @@ contract StakingEscrow is Issuer {
                 uint16 periods = _periods[j];
                 require(value >= minAllowableLockedTokens && periods >= minLockedPeriods);
                 info.value = info.value.add(value);
-                info.subStakes.push(SubStakeInfo(nextPeriod, 0, periods, value));
+                info.subStakes.push(SubStakeInfo(nextPeriod, 0, periods, uint96(value)));
                 sumValue = sumValue.add(value);
                 emit Deposited(staker, value, periods);
                 emit Locked(staker, value, nextPeriod, periods);
@@ -749,12 +749,12 @@ contract StakingEscrow is Issuer {
                 subStake.firstPeriod = _firstPeriod;
                 subStake.lastPeriod = _lastPeriod;
                 subStake.periods = _periods;
-                subStake.lockedValue = _lockedValue;
+                subStake.lockedValue = uint96(_lockedValue);
                 return;
             }
         }
         require(_info.subStakes.length < MAX_SUB_STAKES);
-        _info.subStakes.push(SubStakeInfo(_firstPeriod, _lastPeriod, _periods, _lockedValue));
+        _info.subStakes.push(SubStakeInfo(_firstPeriod, _lastPeriod, _periods, uint96(_lockedValue)));
     }
 
     /**
@@ -773,7 +773,7 @@ contract StakingEscrow is Issuer {
         require(lastPeriod > currentPeriod, "The sub stake must active at least in the next period");
 
         uint256 oldValue = subStake.lockedValue;
-        subStake.lockedValue = oldValue.sub(_newValue);
+        subStake.lockedValue = uint96(oldValue.sub(_newValue));
         require(subStake.lockedValue >= minAllowableLockedTokens);
         uint16 requestedPeriods = subStake.periods.add16(_periods);
         saveSubStake(info, subStake.firstPeriod, 0, requestedPeriods, _newValue);
@@ -976,7 +976,7 @@ contract StakingEscrow is Issuer {
                     lastPeriod.sub16(mintingPeriod));
                 reward += subStakeReward;
                 if (!_info.reStakeDisabled) {
-                    subStake.lockedValue += subStakeReward;
+                    subStake.lockedValue += uint96(subStakeReward);
                 }
             }
         }
@@ -1153,7 +1153,7 @@ contract StakingEscrow is Issuer {
             }
             uint256 appliedPenalty = _penalty;
             if (_penalty < shortestSubStake.lockedValue) {
-                shortestSubStake.lockedValue -= _penalty;
+                shortestSubStake.lockedValue -= uint96(_penalty);
                 saveOldSubStake(_info, shortestSubStake.firstPeriod, _penalty, _decreasePeriod);
                 _penalty = 0;
             } else {
@@ -1252,7 +1252,7 @@ contract StakingEscrow is Issuer {
                 (crossConfirmedPeriod2 ==
                 (oldConfirmedPeriod2 && _info.confirmedPeriod2 >= subStake.firstPeriod))))
             {
-                subStake.lockedValue += _lockedValue;
+                subStake.lockedValue += uint96(_lockedValue);
                 return;
             }
         }
